Add profile menu skeleton
